requirements.txt
dateutils==0.6.6
flickrapi==2.2.1
markovify==0.6.0
Mastodon.py==1.0.6
nltk==3.2.2
oauthlib==2.0.2
olefile==0.44
Pillow==4.1.0
python-dateutil==2.6.0
pytz==2017.2
requests==2.13.0
requests-oauthlib==0.8.0
requests-toolbelt==0.7.1
six==1.10.0
Unidecode==0.4.20
